(design) One socket for each socket on IRC proxy
In the previous commit I wanted to Replayable IRC
to use a shared ws connection to the IRC proxy.
This was a premature optimization trying to
reduce the number of websocket clients to handle
for the IRC proxy.

This optimziation should be either solved in the
ws library used. Which it may already do.
Or be implemented as a auth-mechanism in a layer
above Replayable IRC.