(logapi) The save proxy sends json to the logapi-service
The save-proxy now sends logs to the logapi instead
of directly saving data to the database. This might
be an unecessary abstraction. However I decided to
go for it to check how could split the system into
many different services.

Another concern that come to mind is how I initially
decided to structure the IRC logs. Right now every
user will share the same document for each unique
channel. Im starting to think that was I bad idea.

In the future I will be restructuring the format,
and I do think I will be saving a document for each
connection. That is a user can hold many different
persistent connections. Where each connection owns
a document linking to irc-log documents.

Hopefully, I will draft a detailed description of
the design later.